+++ tests/testthat/test-geom_textpath.R
@@ -16,6 +16,12 @@ test_that("Text path grob has correct types", {
   grobs <- layer_grob(p)[[1]]
 
   expect_s3_class(grobs, "gTree")
+  expect_s3_class(grobs, "textpath")
+  expect_length(grobs$children, 0)
+
+  grobs <- makeContent(grobs)
+
+  expect_length(grobs$children, 2)
   expect_s3_class(grobs$children[[1]], "polyline")
   expect_s3_class(grobs$children[[2]], "text")
 })