+++ NEWS.md
@@ -3,3 +3,5 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Plot text atop curve for readability purposes.
 * Support for `lineend`, `linejoin`, `linemitre` parameters.
+* Letter angles should now be stable for regardless of aspect ratios and 
+  recomputed when the plot device is resized (#6)